requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
mysqlclient>=2.2.0

# Django Channels para WebSockets
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.1.0

# Historial de cambios
django-simple-history>=3.4.0
